Non compile

+++ src/main/java/com/magmaguy/elitemobs/items/ScalableItemConstructor.java
@@ -73,7 +73,7 @@ public class ScalableItemConstructor {
             else
                 secondaryEnchantmentList.put(enchantment, enchantmentsList.get(enchantment));
 
-        int secondaryEnchantmentPool = ThreadLocalRandom.current().nextInt((int) Math.ceil(itemTier / 2));
+        int secondaryEnchantmentPool = ThreadLocalRandom.current().nextInt((int) Math.ceil((double) itemTier / 2));
 
         for (int i = 0; i < secondaryEnchantmentPool; i++) {
